+++ src/main/scala/io/viash/Main.scala
@@ -45,6 +45,8 @@ object Main {
   val version: String = if (pkg.getImplementationVersion != null) pkg.getImplementationVersion else "test"
 
   val viashHome = Paths.get(sys.env.getOrElse("VIASH_HOME", sys.env("HOME") + "/.viash"))
+  val sysEnvOverride = scala.collection.mutable.Map.empty[String, String]
+  def sysEnvGet(key: String): Option[String] = sysEnvOverride.get(key) orElse sys.env.get(key)
 
   /**
     * Viash main
@@ -312,11 +314,24 @@ object Main {
         0
       case List(cli.export, cli.export.cli_schema) =>
         val output = cli.export.cli_schema.output.toOption.map(Paths.get(_))
-        ViashExport.exportCLISchema(output)
+        ViashExport.exportCLISchema(
+          output,
+          format = cli.export.cli_schema.format()
+        )
         0
       case List(cli.export, cli.export.config_schema) =>
         val output = cli.export.config_schema.output.toOption.map(Paths.get(_))
-        ViashExport.exportConfigSchema(output)
+        ViashExport.exportConfigSchema(
+          output,
+          format = cli.export.config_schema.format()
+        )
+        0
+      case List(cli.export, cli.export.json_schema) =>
+        val output = cli.export.json_schema.output.toOption.map(Paths.get(_))
+        ViashExport.exportJsonSchema(
+          output,
+          format = cli.export.json_schema.format()
+        )
         0
       case List(cli.export, cli.export.resource) =>
         val output = cli.export.resource.output.toOption.map(Paths.get(_))
@@ -458,8 +473,8 @@ object Main {
     */
   def detectVersion(workingDir: Option[Path]): Option[String] = {
     // if VIASH_VERSION is defined, use that
-    if (sys.env.get("VIASH_VERSION").isDefined) {
-      sys.env.get("VIASH_VERSION")
+    if (sysEnvGet("VIASH_VERSION").isDefined) {
+      sysEnvGet("VIASH_VERSION")
     } else {
       // else look for project file in working dir
       // and try to read as json