+++ src/main/scala/io/viash/cli/CLIConf.scala
@@ -422,12 +422,19 @@ class CLIConf(arguments: Seq[String]) extends ScallopConf(arguments) {
       banner(
         "viash export cli_schema",
         """Export the schema of the Viash CLI as a JSON""".stripMargin,
-        """viash export cli_schema [--output file.json]""".stripMargin
+        """viash export cli_schema [--output file.json] [--format json]""".stripMargin
       )
       val output = registerOpt[String](
         name = "output",
         default = None,
         descr = "Destination path"
+      )
+      val format = registerChoice(
+        name = "format",
+        short = Some('f'),
+        default = Some("yaml"),
+        choices = List("yaml", "json"),
+        descr = "Which output format to use."
       )
     }
 
@@ -435,18 +442,47 @@ class CLIConf(arguments: Seq[String]) extends ScallopConf(arguments) {
       banner(
         "viash export config_schema",
         """Export the schema of a Viash config as a JSON""".stripMargin,
-        """viash export config_schema [--output file.json]""".stripMargin
+        """viash export config_schema [--output file.json] [--format json]""".stripMargin
       )
       val output = registerOpt[String](
         name = "output",
         default = None,
         descr = "Destination path"
       )
+      val format = registerChoice(
+        name = "format",
+        short = Some('f'),
+        default = Some("yaml"),
+        choices = List("yaml", "json"),
+        descr = "Which output format to use."
+      )
     }
+
+    val json_schema = new DocumentedSubcommand("json_schema") {
+      banner(
+        "viash export json_schema",
+        """Export the json schema to validate a Viash config""".stripMargin,
+        """viash export json_schema [--output file.json] [--format json]""".stripMargin
+      )
+      val output = registerOpt[String](
+        name = "output",
+        default = None,
+        descr = "Destination path"
+      )
+      val format = registerChoice(
+        name = "format",
+        short = Some('f'),
+        default = Some("yaml"),
+        choices = List("yaml", "json"),
+        descr = "Which output format to use."
+      )
+    }
+
 
     addSubcommand(resource)
     addSubcommand(cli_schema)
     addSubcommand(config_schema)
+    addSubcommand(json_schema)
 
     requireSubcommand()