+++ src/main/scala/io/viash/functionality/Functionality.scala
@@ -74,6 +74,7 @@ case class Functionality(
       |    email: tim@far.be
       |""".stripMargin, "yaml")
   @since("Viash 0.3.1")
+  @default("Empty")
   authors: List[Author] = Nil,
   
   @description(
@@ -103,6 +104,7 @@ case class Functionality(
       |    type: string
       |""".stripMargin,
       "yaml")
+  @default("Empty")
   arguments: List[Argument[_]] = Nil,
 
   @description(
@@ -110,7 +112,7 @@ case class Functionality(
       |
       | - `name: foo`, the name of the argument group. 
       | - `description: Description of foo`, a description of the argument group. Multiline descriptions are supported.
-      | - `arguments: [arg1, arg2, ...]`, list of the arguments names.
+      | - `arguments: [arg1, arg2, ...]`, list of the arguments.
       |
       |""".stripMargin)
   @example(
@@ -154,6 +156,7 @@ case class Functionality(
       "bash",
       "This results in the following output when calling the component with the `--help` argument:")
   @since("Viash 0.5.14")
+  @default("Empty")
   argument_groups: List[ArgumentGroup] = Nil,
 
   @description(
@@ -175,6 +178,7 @@ case class Functionality(
       |    path: resource1.txt
       |""".stripMargin,
       "yaml")
+  @default("Empty")
   resources: List[Resource] = Nil,
 
   @description("A description of the component. This will be displayed with `--help`.")
@@ -200,6 +204,7 @@ case class Functionality(
       |  - path: resource1.txt
       |""".stripMargin,
       "yaml")
+  @default("Empty")
   test_resources: List[Resource] = Nil,
 
   @description("Structured information. Can be any shape: a string, vector, map or even nested map.")
@@ -208,10 +213,12 @@ case class Functionality(
       |  twitter: wizzkid
       |  classes: [ one, two, three ]""".stripMargin, "yaml")
   @since("Viash 0.4.0")
+  @default("Empty")
   info: Json = Json.Null,
 
   @description("Allows setting a component to active, deprecated or disabled.")
   @since("Viash 0.6.0")
+  @default("Enabled")
   status: Status = Status.Enabled,
   
   @description(
@@ -226,6 +233,7 @@ case class Functionality(
       |""".stripMargin,
       "yaml")
   @since("Viash 0.6.0")
+  @default("Empty")
   requirements: ComputationalRequirements = ComputationalRequirements(),
 
   // The variables below are for internal use and shouldn't be publicly documented