+++ src/main/scala/io/viash/schemas/CollectedSchemas.scala
@@ -101,11 +101,13 @@ object CollectedSchemas {
     "config" -> Map(
       "config"                    -> getMembers[Config](),
       "project"                   -> getMembers[ViashProject](),
+      "info"                      -> getMembers[Info](),
     ),
     "functionality" -> Map(
       "functionality"             -> getMembers[Functionality](),
       "author"                    -> getMembers[Author](),
       "computationalRequirements" -> getMembers[ComputationalRequirements](),
+      "argumentGroup"             -> getMembers[ArgumentGroup](),
     ),
     "platforms" -> Map(
       "platform"                  -> getMembers[Platform](),
@@ -158,7 +160,7 @@ object CollectedSchemas {
   private def trimTypeName(s: String) = {
     // first: io.viash.helpers.data_structures.OneOrMore[String] -> OneOrMore[String]
     // second: List[io.viash.platforms.requirements.Requirements] -> List[Requirements]
-    s.replaceAll("^(\\w*\\.)*", "").replaceAll("""(\w*)\[[\w\.]*?([\w,]*)(\[_\])?\]""", "$1 of $2")
+    s.replaceAll("^(\\w*\\.)*", "").replaceAll("""(\w*)\[[\w\.]*?([\w,]*)(\[_\])?\]""", "$1[$2]")
   }
 
   private def annotationsOf(members: (Map[String,List[MemberInfo]]), classes: List[Symbol]) = {
@@ -172,17 +174,17 @@ object CollectedSchemas {
     val annThis = ("__this__", classes.head.name.toString(), classes.head.annotations, "", 0, classes.map(_.fullName))
     val allAnnotations = annThis :: annMembers.toList
     allAnnotations
-      .map({case (a, b, c, d, e, f) => (a, trimTypeName(b), f, c)})  // TODO this ignores where the annotation was defined, ie. top level class or super class
+      .map({case (name, tpe, annotations, d, e, hierarchy) => (name, trimTypeName(tpe), hierarchy, annotations)})  // TODO this ignores where the annotation was defined, ie. top level class or super class
   }
 
   private val getSchema = (t: (Map[String,List[MemberInfo]], List[Symbol])) => t match {
     case (members, classes) => {
-      annotationsOf(members, classes).flatMap{ case (a, b, c, d) => ParameterSchema(a, b, c, d) }
+      annotationsOf(members, classes).flatMap{ case (name, tpe, hierarchy, annotations) => ParameterSchema(name, tpe, hierarchy, annotations) }
     }
   }
 
   // Main call for documentation output
-  private lazy val data = CollectedSchemas(
+  lazy val data = CollectedSchemas(
       config = schemaClassMap.get("config").get.map{ case(k, v) => (k, getSchema(v))},
       functionality = schemaClassMap.get("functionality").get.map{ case(k, v) => (k, getSchema(v))},
       platforms = schemaClassMap.get("platforms").get.map{ case(k, v) => (k, getSchema(v))},